refactor: consolidate all type definitions into types.ts (#431)
- Move all component prop interfaces and handle types from individual
  component files into the central src/types.ts
- Components now import their types from types.ts instead of defining
  them locally
- Re-export all type definitions from index.ts for consumer access
- No behavioral changes — only import paths changed
