+++ src/index.ts
@@ -37,4 +37,18 @@ export type {
   SortByShape,
   SortState,
   TableComponents,
+  AutoResizerProps,
+  BaseTableProps,
+  BaseTableHandle,
+  ColumnResizerProps,
+  ExpandIconProps,
+  GridTableProps,
+  GridTableHandle,
+  SortIndicatorProps,
+  TableCellProps,
+  TableHeaderProps,
+  TableHeaderHandle,
+  TableHeaderCellProps,
+  TableHeaderRowProps,
+  TableRowProps,
 } from './types';
